feat: Move BreedPicture cell to UI core
