+++ DogBreed/Packages/DogUI/Sources/DogUI/BreedPictures/DogUI.BreedPictures.swift
@@ -1,0 +1,13 @@
+//
+//  File.swift
+//  DogUIPackage
+//
+//  Created by Matej on 12. 6. 25.
+//
+
+import Foundation
+
+public extension DogUI {
+
+    enum BreedPictures {}
+}
