+++ DogBreed/Packages/DogUI/Sources/DogUI/DogUI.swift
@@ -1,0 +1,10 @@
+//
+//  File.swift
+//  DogUIPackage
+//
+//  Created by Matej on 12. 6. 25.
+//
+
+import Foundation
+
+public enum DogUI {}
