+++ DogBreed/Packages/DogUI/Sources/DogUI/Extension/UICollectionViewCell+Identifier.swift
@@ -8,8 +8,7 @@
 
 import UIKit
 
-extension UICollectionViewCell {
+public extension UICollectionViewCell {
 
     class var identifier: String { String(describing: "\(self)") }
-
 }
